Gordon Resting-State Cortical Parcellation

Brain atlas for the Gordon cortical parcellation with 329 regions based on resting-state functional connectivity boundaries. Contains 2D polygon geometry for ggseg::geom_brain().

gordon()

Value

A ggseg.formats::ggseg_atlas object (cortical).

References

Gordon EM, Laumann TO, Adeyemo B, Huckins JF, Kelley WM, Petersen SE (2016). Generation and Evaluation of a Cortical Area Parcellation from Resting-State Correlations. Cerebral Cortex, 26(1):288-303. doi:10.1093/cercor/bhu239

Examples

gordon()
#> 
#> ── gordon ggseg atlas ──────────────────────────────────────────────────────────
#> Type: cortical
#> Regions: 329
#> Hemispheres: left, right
#> Views: lateral, medial
#> Palette: ✔
#> Rendering: ✔ ggseg
#> ✔ ggseg3d (vertices)
#> ────────────────────────────────────────────────────────────────────────────────
#> # A tibble: 329 × 3
#>     hemi  region            label                 
#>     <chr> <chr>             <chr>                 
#>   1 left  None_4            lh_L_None_4           
#>   2 left  None_5            lh_L_None_5           
#>   3 left  None_6            lh_L_None_6           
#>   4 left  None_7            lh_L_None_7           
#>   5 left  None_8            lh_L_None_8           
#>   6 left  None_9            lh_L_None_9           
#>   7 left  None_14           lh_L_None_14          
#>   8 left  None_15           lh_L_None_15          
#>   9 left  None_16           lh_L_None_16          
#>  10 left  None_21           lh_L_None_21          
#>  11 left  Default_2         lh_L_Default_2        
#>  12 left  Default_3         lh_L_Default_3        
#>  13 left  Default_6         lh_L_Default_6        
#>  14 left  Default_7         lh_L_Default_7        
#>  15 left  Default_8         lh_L_Default_8        
#>  16 left  Default_11        lh_L_Default_11       
#>  17 left  Default_12        lh_L_Default_12       
#>  18 left  Default_13        lh_L_Default_13       
#>  19 left  Default_14        lh_L_Default_14       
#>  20 left  Default_16        lh_L_Default_16       
#>  21 left  Default_18        lh_L_Default_18       
#>  22 left  Default_19        lh_L_Default_19       
#>  23 left  Default_20        lh_L_Default_20       
#>  24 left  FrontoParietal_1  lh_L_FrontoParietal_1 
#>  25 left  FrontoParietal_2  lh_L_FrontoParietal_2 
#>  26 left  FrontoParietal_4  lh_L_FrontoParietal_4 
#>  27 left  FrontoParietal_5  lh_L_FrontoParietal_5 
#>  28 left  FrontoParietal_6  lh_L_FrontoParietal_6 
#>  29 left  FrontoParietal_7  lh_L_FrontoParietal_7 
#>  30 left  FrontoParietal_8  lh_L_FrontoParietal_8 
#>  31 left  FrontoParietal_9  lh_L_FrontoParietal_9 
#>  32 left  Salience_2        lh_L_Salience_2       
#>  33 left  CinguloOperc_6    lh_L_CinguloOperc_6   
#>  34 left  CinguloOperc_7    lh_L_CinguloOperc_7   
#>  35 left  CinguloOperc_8    lh_L_CinguloOperc_8   
#>  36 left  CinguloOperc_9    lh_L_CinguloOperc_9   
#>  37 left  CinguloOperc_10   lh_L_CinguloOperc_10  
#>  38 left  CinguloOperc_11   lh_L_CinguloOperc_11  
#>  39 left  CinguloOperc_12   lh_L_CinguloOperc_12  
#>  40 left  CinguloOperc_13   lh_L_CinguloOperc_13  
#>  41 left  CinguloOperc_14   lh_L_CinguloOperc_14  
#>  42 left  CinguloOperc_15   lh_L_CinguloOperc_15  
#>  43 left  CinguloOperc_16   lh_L_CinguloOperc_16  
#>  44 left  CinguloOperc_17   lh_L_CinguloOperc_17  
#>  45 left  CinguloOperc_18   lh_L_CinguloOperc_18  
#>  46 left  CinguloOperc_19   lh_L_CinguloOperc_19  
#>  47 left  CinguloOperc_20   lh_L_CinguloOperc_20  
#>  48 left  DorsalAttn_1      lh_L_DorsalAttn_1     
#>  49 left  DorsalAttn_2      lh_L_DorsalAttn_2     
#>  50 left  DorsalAttn_3      lh_L_DorsalAttn_3     
#>  51 left  DorsalAttn_4      lh_L_DorsalAttn_4     
#>  52 left  DorsalAttn_5      lh_L_DorsalAttn_5     
#>  53 left  DorsalAttn_6      lh_L_DorsalAttn_6     
#>  54 left  DorsalAttn_7      lh_L_DorsalAttn_7     
#>  55 left  DorsalAttn_8      lh_L_DorsalAttn_8     
#>  56 left  DorsalAttn_9      lh_L_DorsalAttn_9     
#>  57 left  DorsalAttn_10     lh_L_DorsalAttn_10    
#>  58 left  DorsalAttn_13     lh_L_DorsalAttn_13    
#>  59 left  DorsalAttn_14     lh_L_DorsalAttn_14    
#>  60 left  DorsalAttn_15     lh_L_DorsalAttn_15    
#>  61 left  DorsalAttn_16     lh_L_DorsalAttn_16    
#>  62 left  DorsalAttn_17     lh_L_DorsalAttn_17    
#>  63 left  DorsalAttn_18     lh_L_DorsalAttn_18    
#>  64 left  DorsalAttn_19     lh_L_DorsalAttn_19    
#>  65 left  VentralAttn_2     lh_L_VentralAttn_2    
#>  66 left  VentralAttn_3     lh_L_VentralAttn_3    
#>  67 left  VentralAttn_4     lh_L_VentralAttn_4    
#>  68 left  VentralAttn_5     lh_L_VentralAttn_5    
#>  69 left  VentralAttn_6     lh_L_VentralAttn_6    
#>  70 left  VentralAttn_7     lh_L_VentralAttn_7    
#>  71 left  VentralAttn_8     lh_L_VentralAttn_8    
#>  72 left  VentralAttn_9     lh_L_VentralAttn_9    
#>  73 left  VentralAttn_10    lh_L_VentralAttn_10   
#>  74 left  VentralAttn_11    lh_L_VentralAttn_11   
#>  75 left  Visual_1          lh_L_Visual_1         
#>  76 left  Visual_8          lh_L_Visual_8         
#>  77 left  Visual_9          lh_L_Visual_9         
#>  78 left  Visual_10         lh_L_Visual_10        
#>  79 left  Visual_16         lh_L_Visual_16        
#>  80 left  Visual_17         lh_L_Visual_17        
#>  81 left  Visual_18         lh_L_Visual_18        
#>  82 left  SMhand_1          lh_L_SMhand_1         
#>  83 left  SMhand_3          lh_L_SMhand_3         
#>  84 left  SMhand_4          lh_L_SMhand_4         
#>  85 left  SMhand_9          lh_L_SMhand_9         
#>  86 left  SMhand_10         lh_L_SMhand_10        
#>  87 left  SMhand_11         lh_L_SMhand_11        
#>  88 left  SMhand_12         lh_L_SMhand_12        
#>  89 left  SMhand_13         lh_L_SMhand_13        
#>  90 left  SMhand_14         lh_L_SMhand_14        
#>  91 left  SMhand_15         lh_L_SMhand_15        
#>  92 left  SMhand_16         lh_L_SMhand_16        
#>  93 left  SMhand_17         lh_L_SMhand_17        
#>  94 left  SMhand_18         lh_L_SMhand_18        
#>  95 left  SMmouth_1         lh_L_SMmouth_1        
#>  96 left  SMmouth_2         lh_L_SMmouth_2        
#>  97 left  SMmouth_3         lh_L_SMmouth_3        
#>  98 left  SMmouth_4         lh_L_SMmouth_4        
#>  99 left  Auditory_1        lh_L_Auditory_1       
#> 100 left  Auditory_2        lh_L_Auditory_2       
#> 101 left  Auditory_3        lh_L_Auditory_3       
#> 102 left  Auditory_4        lh_L_Auditory_4       
#> 103 left  Auditory_5        lh_L_Auditory_5       
#> 104 left  Auditory_6        lh_L_Auditory_6       
#> 105 left  Auditory_7        lh_L_Auditory_7       
#> 106 left  Auditory_8        lh_L_Auditory_8       
#> 107 left  Auditory_9        lh_L_Auditory_9       
#> 108 left  Auditory_10       lh_L_Auditory_10      
#> 109 left  Auditory_11       lh_L_Auditory_11      
#> 110 left  Auditory_12       lh_L_Auditory_12      
#> 111 left  None_1            lh_L_None_1           
#> 112 left  None_2            lh_L_None_2           
#> 113 left  None_3            lh_L_None_3           
#> 114 left  None_11           lh_L_None_11          
#> 115 left  None_12           lh_L_None_12          
#> 116 left  None_13           lh_L_None_13          
#> 117 left  None_17           lh_L_None_17          
#> 118 left  None_18           lh_L_None_18          
#> 119 left  None_19           lh_L_None_19          
#> 120 left  None_20           lh_L_None_20          
#> 121 left  Default_1         lh_L_Default_1        
#> 122 left  Default_4         lh_L_Default_4        
#> 123 left  Default_5         lh_L_Default_5        
#> 124 left  Default_9         lh_L_Default_9        
#> 125 left  Default_10        lh_L_Default_10       
#> 126 left  Default_15        lh_L_Default_15       
#> 127 left  Default_17        lh_L_Default_17       
#> 128 left  ParietoOccip_1    lh_L_ParietoOccip_1   
#> 129 left  ParietoOccip_2    lh_L_ParietoOccip_2   
#> 130 left  ParietoOccip_3    lh_L_ParietoOccip_3   
#> 131 left  ParietoOccip_4    lh_L_ParietoOccip_4   
#> 132 left  FrontoParietal_3  lh_L_FrontoParietal_3 
#> 133 left  Salience_1        lh_L_Salience_1       
#> 134 left  CinguloOperc_1    lh_L_CinguloOperc_1   
#> 135 left  CinguloOperc_2    lh_L_CinguloOperc_2   
#> 136 left  CinguloOperc_3    lh_L_CinguloOperc_3   
#> 137 left  CinguloOperc_4    lh_L_CinguloOperc_4   
#> 138 left  CinguloOperc_5    lh_L_CinguloOperc_5   
#> 139 left  MedialParietal_1  lh_L_MedialParietal_1 
#> 140 left  MedialParietal_2  lh_L_MedialParietal_2 
#> 141 left  MedialParietal_3  lh_L_MedialParietal_3 
#> 142 left  DorsalAttn_11     lh_L_DorsalAttn_11    
#> 143 left  DorsalAttn_12     lh_L_DorsalAttn_12    
#> 144 left  VentralAttn_1     lh_L_VentralAttn_1    
#> 145 left  Visual_2          lh_L_Visual_2         
#> 146 left  Visual_3          lh_L_Visual_3         
#> 147 left  Visual_4          lh_L_Visual_4         
#> 148 left  Visual_5          lh_L_Visual_5         
#> 149 left  Visual_6          lh_L_Visual_6         
#> 150 left  Visual_7          lh_L_Visual_7         
#> 151 left  Visual_11         lh_L_Visual_11        
#> 152 left  Visual_12         lh_L_Visual_12        
#> 153 left  Visual_13         lh_L_Visual_13        
#> 154 left  Visual_14         lh_L_Visual_14        
#> 155 left  Visual_15         lh_L_Visual_15        
#> 156 left  SMhand_2          lh_L_SMhand_2         
#> 157 left  SMhand_5          lh_L_SMhand_5         
#> 158 left  SMhand_6          lh_L_SMhand_6         
#> 159 left  SMhand_7          lh_L_SMhand_7         
#> 160 left  SMhand_8          lh_L_SMhand_8         
#> 161 right None_22           rh_R_None_22          
#> 162 right None_23           rh_R_None_23          
#> 163 right None_24           rh_R_None_24          
#> 164 right None_25           rh_R_None_25          
#> 165 right None_31           rh_R_None_31          
#> 166 right None_32           rh_R_None_32          
#> 167 right None_33           rh_R_None_33          
#> 168 right None_37           rh_R_None_37          
#> 169 right None_38           rh_R_None_38          
#> 170 right None_39           rh_R_None_39          
#> 171 right None_42           rh_R_None_42          
#> 172 right None_43           rh_R_None_43          
#> 173 right None_44           rh_R_None_44          
#> 174 right None_45           rh_R_None_45          
#> 175 right None_46           rh_R_None_46          
#> 176 right Default_21        rh_R_Default_21       
#> 177 right Default_22        rh_R_Default_22       
#> 178 right Default_23        rh_R_Default_23       
#> 179 right Default_24        rh_R_Default_24       
#> 180 right Default_28        rh_R_Default_28       
#> 181 right Default_30        rh_R_Default_30       
#> 182 right Default_31        rh_R_Default_31       
#> 183 right Default_36        rh_R_Default_36       
#> 184 right Default_37        rh_R_Default_37       
#> 185 right Default_38        rh_R_Default_38       
#> 186 right Default_39        rh_R_Default_39       
#> 187 right ParietoOccip_5    rh_R_ParietoOccip_5   
#> 188 right ParietoOccip_6    rh_R_ParietoOccip_6   
#> 189 right ParietoOccip_7    rh_R_ParietoOccip_7   
#> 190 right ParietoOccip_8    rh_R_ParietoOccip_8   
#> 191 right FrontoParietal_13 rh_R_FrontoParietal_13
#> 192 right Salience_3        rh_R_Salience_3       
#> 193 right CinguloOperc_21   rh_R_CinguloOperc_21  
#> 194 right CinguloOperc_22   rh_R_CinguloOperc_22  
#> 195 right CinguloOperc_23   rh_R_CinguloOperc_23  
#> 196 right CinguloOperc_24   rh_R_CinguloOperc_24  
#> 197 right CinguloOperc_25   rh_R_CinguloOperc_25  
#> 198 right CinguloOperc_26   rh_R_CinguloOperc_26  
#> 199 right CinguloOperc_27   rh_R_CinguloOperc_27  
#> 200 right MedialParietal_4  rh_R_MedialParietal_4 
#> 201 right MedialParietal_5  rh_R_MedialParietal_5 
#> 202 right DorsalAttn_20     rh_R_DorsalAttn_20    
#> 203 right DorsalAttn_27     rh_R_DorsalAttn_27    
#> 204 right Visual_20         rh_R_Visual_20        
#> 205 right Visual_21         rh_R_Visual_21        
#> 206 right Visual_22         rh_R_Visual_22        
#> 207 right Visual_23         rh_R_Visual_23        
#> 208 right Visual_25         rh_R_Visual_25        
#> 209 right Visual_26         rh_R_Visual_26        
#> 210 right Visual_32         rh_R_Visual_32        
#> 211 right Visual_33         rh_R_Visual_33        
#> 212 right Visual_34         rh_R_Visual_34        
#> 213 right Visual_35         rh_R_Visual_35        
#> 214 right Visual_36         rh_R_Visual_36        
#> 215 right Visual_37         rh_R_Visual_37        
#> 216 right Visual_38         rh_R_Visual_38        
#> 217 right Visual_39         rh_R_Visual_39        
#> 218 right SMhand_19         rh_R_SMhand_19        
#> 219 right SMhand_20         rh_R_SMhand_20        
#> 220 right SMhand_21         rh_R_SMhand_21        
#> 221 right SMhand_22         rh_R_SMhand_22        
#> 222 right SMhand_23         rh_R_SMhand_23        
#> 223 right SMhand_24         rh_R_SMhand_24        
#> 224 right None_26           rh_R_None_26          
#> 225 right None_28           rh_R_None_28          
#> 226 right None_29           rh_R_None_29          
#> 227 right None_34           rh_R_None_34          
#> 228 right None_35           rh_R_None_35          
#> 229 right None_36           rh_R_None_36          
#> 230 right None_40           rh_R_None_40          
#> 231 right None_41           rh_R_None_41          
#> 232 right Default_25        rh_R_Default_25       
#> 233 right Default_26        rh_R_Default_26       
#> 234 right Default_27        rh_R_Default_27       
#> 235 right Default_29        rh_R_Default_29       
#> 236 right Default_32        rh_R_Default_32       
#> 237 right Default_33        rh_R_Default_33       
#> 238 right Default_34        rh_R_Default_34       
#> 239 right Default_35        rh_R_Default_35       
#> 240 right Default_40        rh_R_Default_40       
#> 241 right Default_41        rh_R_Default_41       
#> 242 right FrontoParietal_10 rh_R_FrontoParietal_10
#> 243 right FrontoParietal_11 rh_R_FrontoParietal_11
#> 244 right FrontoParietal_12 rh_R_FrontoParietal_12
#> 245 right FrontoParietal_14 rh_R_FrontoParietal_14
#> 246 right FrontoParietal_15 rh_R_FrontoParietal_15
#> 247 right FrontoParietal_16 rh_R_FrontoParietal_16
#> 248 right FrontoParietal_17 rh_R_FrontoParietal_17
#> 249 right FrontoParietal_18 rh_R_FrontoParietal_18
#> 250 right FrontoParietal_19 rh_R_FrontoParietal_19
#> 251 right FrontoParietal_20 rh_R_FrontoParietal_20
#> 252 right FrontoParietal_21 rh_R_FrontoParietal_21
#> 253 right FrontoParietal_22 rh_R_FrontoParietal_22
#> 254 right FrontoParietal_23 rh_R_FrontoParietal_23
#> 255 right FrontoParietal_24 rh_R_FrontoParietal_24
#> 256 right Salience_4        rh_R_Salience_4       
#> 257 right CinguloOperc_28   rh_R_CinguloOperc_28  
#> 258 right CinguloOperc_29   rh_R_CinguloOperc_29  
#> 259 right CinguloOperc_30   rh_R_CinguloOperc_30  
#> 260 right CinguloOperc_31   rh_R_CinguloOperc_31  
#> 261 right CinguloOperc_32   rh_R_CinguloOperc_32  
#> 262 right CinguloOperc_33   rh_R_CinguloOperc_33  
#> 263 right CinguloOperc_34   rh_R_CinguloOperc_34  
#> 264 right CinguloOperc_35   rh_R_CinguloOperc_35  
#> 265 right CinguloOperc_36   rh_R_CinguloOperc_36  
#> 266 right CinguloOperc_37   rh_R_CinguloOperc_37  
#> 267 right CinguloOperc_38   rh_R_CinguloOperc_38  
#> 268 right CinguloOperc_39   rh_R_CinguloOperc_39  
#> 269 right CinguloOperc_40   rh_R_CinguloOperc_40  
#> 270 right DorsalAttn_21     rh_R_DorsalAttn_21    
#> 271 right DorsalAttn_22     rh_R_DorsalAttn_22    
#> 272 right DorsalAttn_23     rh_R_DorsalAttn_23    
#> 273 right DorsalAttn_24     rh_R_DorsalAttn_24    
#> 274 right DorsalAttn_25     rh_R_DorsalAttn_25    
#> 275 right DorsalAttn_26     rh_R_DorsalAttn_26    
#> 276 right DorsalAttn_28     rh_R_DorsalAttn_28    
#> 277 right DorsalAttn_29     rh_R_DorsalAttn_29    
#> 278 right DorsalAttn_30     rh_R_DorsalAttn_30    
#> 279 right DorsalAttn_31     rh_R_DorsalAttn_31    
#> 280 right DorsalAttn_32     rh_R_DorsalAttn_32    
#> 281 right VentralAttn_12    rh_R_VentralAttn_12   
#> 282 right VentralAttn_13    rh_R_VentralAttn_13   
#> 283 right VentralAttn_14    rh_R_VentralAttn_14   
#> 284 right VentralAttn_15    rh_R_VentralAttn_15   
#> 285 right VentralAttn_16    rh_R_VentralAttn_16   
#> 286 right VentralAttn_17    rh_R_VentralAttn_17   
#> 287 right VentralAttn_18    rh_R_VentralAttn_18   
#> 288 right VentralAttn_19    rh_R_VentralAttn_19   
#> 289 right VentralAttn_20    rh_R_VentralAttn_20   
#> 290 right VentralAttn_21    rh_R_VentralAttn_21   
#> 291 right VentralAttn_22    rh_R_VentralAttn_22   
#> 292 right VentralAttn_23    rh_R_VentralAttn_23   
#> 293 right Visual_19         rh_R_Visual_19        
#> 294 right Visual_24         rh_R_Visual_24        
#> 295 right Visual_27         rh_R_Visual_27        
#> 296 right Visual_28         rh_R_Visual_28        
#> 297 right Visual_29         rh_R_Visual_29        
#> 298 right Visual_30         rh_R_Visual_30        
#> 299 right Visual_31         rh_R_Visual_31        
#> 300 right SMhand_25         rh_R_SMhand_25        
#> 301 right SMhand_26         rh_R_SMhand_26        
#> 302 right SMhand_27         rh_R_SMhand_27        
#> 303 right SMhand_28         rh_R_SMhand_28        
#> 304 right SMhand_29         rh_R_SMhand_29        
#> 305 right SMhand_30         rh_R_SMhand_30        
#> 306 right SMhand_31         rh_R_SMhand_31        
#> 307 right SMhand_32         rh_R_SMhand_32        
#> 308 right SMhand_33         rh_R_SMhand_33        
#> 309 right SMhand_34         rh_R_SMhand_34        
#> 310 right SMhand_35         rh_R_SMhand_35        
#> 311 right SMhand_36         rh_R_SMhand_36        
#> 312 right SMhand_37         rh_R_SMhand_37        
#> 313 right SMhand_38         rh_R_SMhand_38        
#> 314 right SMmouth_5         rh_R_SMmouth_5        
#> 315 right SMmouth_6         rh_R_SMmouth_6        
#> 316 right SMmouth_7         rh_R_SMmouth_7        
#> 317 right SMmouth_8         rh_R_SMmouth_8        
#> 318 right Auditory_13       rh_R_Auditory_13      
#> 319 right Auditory_14       rh_R_Auditory_14      
#> 320 right Auditory_15       rh_R_Auditory_15      
#> 321 right Auditory_16       rh_R_Auditory_16      
#> 322 right Auditory_17       rh_R_Auditory_17      
#> 323 right Auditory_18       rh_R_Auditory_18      
#> 324 right Auditory_19       rh_R_Auditory_19      
#> 325 right Auditory_20       rh_R_Auditory_20      
#> 326 right Auditory_21       rh_R_Auditory_21      
#> 327 right Auditory_22       rh_R_Auditory_22      
#> 328 right Auditory_23       rh_R_Auditory_23      
#> 329 right Auditory_24       rh_R_Auditory_24